+++ changelog.md
@@ -7,6 +7,8 @@
 ##### New Features
 
 * Added basic client method to run hosting capacity work packages
+* Added basic client method to cancel hosting capacity work packages
+* Added basic client method to request hosting capacity work packages progress
 
 ##### Enhancements